Ensure custom matplotlib.text filter is added with lazy imports
